requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
execnet==1.9.0
factory-boy==3.2.0
Faker==9.5.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
kafka-python==2.0.2
lxml==4.6.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytest-forked==1.3.0
pytest-mock==3.6.1
pytest-sugar==0.9.4
pytest-xdist==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
slack-sdk==3.11.2
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
